requirements.txt
Flask==2.2.2
google-cloud-storage==2.6.0